+++ src/main/java/Components/HeuristicProposal.java
@@ -1,8 +1,8 @@
 package Components;
 
 public class HeuristicProposal {
-    int h;
+   /* int h;
     Agent a;
-    int taskID;
+    int taskID;*/
 
 }
