+++ src/main/java/Components/State/Wall.java
@@ -1,13 +1,16 @@
 package Components.State;
 
+
+//THIS IS PROBABLY REDUNDANT
+//TODO: MAYBE REFACTOR TO BOOLEAN VALUE IN TILE
 public class Wall {
 
     private final int column;
     private final int row;
 
-    public Wall(final int column, final int row) {
-        this.column = column;
+    public Wall(final int row, final int column) {
         this.row = row;
+        this.column = column;
     }
